+++ public/CHANGELOG.md
@@ -7,3 +7,4 @@
 - 添加原生属性保存功能 | Add native attribute save function.
 - 添加书签列表查询功能 | Add bookmark list query function.
 - 添加浏览位置属性编辑功能 | Add browser location attribute edit function.
+- 添加自定义属性编辑功能 | Add custom attributes edit function.